+++ src/components/accordion/accordion-content.tsx
@@ -1,3 +1,4 @@
+import classNames from "classnames";
 import React, { PureComponent } from "react";
 import { CSSTransition } from "react-transition-group";
 import IDiv from "../../interfaces/div";
@@ -7,7 +8,8 @@ import { animationTypes } from "../../types/modal";
 export default class Content extends PureComponent<IAccordionContentProps> {
 
   public render() {
-    const { expanded, children } = this.props;
+    const { expanded, className, children } = this.props;
+    const accordionContentClasses = classNames("q-accordion-content", className);
     return (
       <CSSTransition
         in={expanded}
@@ -15,7 +17,7 @@ export default class Content extends PureComponent<IAccordionContentProps> {
         timeout={300}
         classNames="q-slideInDown q-accordion"
       >
-        <div className="q-accordion-content">{children}</div>
+        <div className={accordionContentClasses}>{children}</div>
       </CSSTransition>
     );
   }