+++ src/components/radio/radio.tsx
@@ -2,11 +2,24 @@ import classNames from "classnames";
 import React, { PureComponent } from "react";
 import IInput from "../../interfaces/input";
 import "../../styles/components/_radio.scss";
+import { colorTypes } from "../../types/color";
+import { variantTypes } from "../../types/typography";
 import classNamesDefault from "../../utils/class-names-default";
+import Typography from "../typography";
 
 export default class Radio extends PureComponent<IProps> {
   public render() {
-    const { checked, error, label, type, value, className, ...props } = this.props;
+    const {
+      checked,
+      error,
+      label,
+      labelVariant = "body",
+      labelColor= "black",
+      type,
+      value,
+      className,
+      ...props
+    } = this.props;
     const radioClasses = classNames(
       "q-radio-wrapper",
       classNamesDefault({ error }),
@@ -22,7 +35,11 @@ export default class Radio extends PureComponent<IProps> {
           {...props}
           defaultChecked={checked}
         />
-        {label && <label className={"q-label"} htmlFor={`radio-${value}`}>{label}</label>}
+        {label && <label className={"q-label"} htmlFor={`radio-${value}`}>
+          <Typography variant={labelVariant} color={labelColor} display={"inline"}>
+            {label}
+          </Typography>
+        </label>}
       </div>
     );
   }
@@ -31,6 +48,8 @@ export default class Radio extends PureComponent<IProps> {
 interface IProps extends IInput {
   error?: boolean;
   label?: string;
-  value: string;
+  labelVariant?: variantTypes;
+  labelColor?: colorTypes;
+  value: any;
   className?: string;
 }