+++ src/stories/checkbox.stories.tsx
@@ -31,3 +31,28 @@ stories.add("Default", () => (
     />
   </div>
 ));
+stories.add("Label Size", () => (
+  <div>
+      <CheckBox name={"vehicle1"} value={"1"} label={"I have a bike"} labelVariant={"h1"}/>
+      <br/>
+      <br/>
+      <CheckBox name={"vehicle1"} value={"2"} label={"I have a bike"} labelVariant={"h2"}/>
+      <br/>
+      <br/>
+      <CheckBox name={"vehicle1"} value={"3"} label={"I have a bike"} labelVariant={"h3"}/>
+      <br/>
+      <br/>
+      <CheckBox name={"vehicle1"} value={"4"} label={"I have a bike"} labelVariant={"paragraph"}/>
+      <br/>
+      <br/>
+      <CheckBox name={"vehicle1"} value={"5"} label={"I have a bike"} labelVariant={"smallParagraph"}/>
+      <br/>
+      <br/>
+      <CheckBox name={"vehicle1"} value={"6"} label={"I have a bike"} labelVariant={"body"}/>
+      <br/>
+      <br/>
+      <CheckBox name={"vehicle1"} value={"7"} label={"I have a bike"} labelVariant={"subtitle"}/>
+      <br/>
+      <br/>
+  </div>
+));