+++ src/types/color.ts
@@ -6,4 +6,6 @@ export type colorTypes =
   "dark-gray" |
   "light-gray" |
   "border-gray" |
-  "white";
+  "white" |
+  "yellow" |
+  "zero";