+++ src/types/icon.ts
@@ -1,2 +1,2 @@
-export type sizeTypes = "small" | "large" | "xlarge";
+export type sizeTypes = "large" | "medium" | "small" | "xlarge";
 export type variantTypes = "primary" | "gray";