+++ src/types/rating.ts
@@ -1,0 +1,1 @@
+export type ratingSize = "large" | "medium" | "small";